Make QueueFile an interface
