+++ tape/src/main/java/com/squareup/tape/FileException.java
@@ -6,14 +6,14 @@ import java.io.IOException;
 
 /** Encapsulates an {@link IOException} in an extension of {@link RuntimeException}. */
 public class FileException extends RuntimeException {
-  private final File file;
+  private final QueueFile file;
 
-  public FileException(String message, IOException e, File file) {
+  public FileException(String message, IOException e, QueueFile file) {
     super(message, e);
     this.file = file;
   }
 
-  public File getFile() {
+  public QueueFile getFile() {
     return file;
   }
 }
